Update disable-encryption-by-default-97-higher-task.adoc
The -vserver attribute in the command is not used by "options".
I clarified that this option applies globally - across the entire cluster.
I also removed the "see the man page" sentence, because the "options" CLI command is a hidden (!) command. There is no man page for "options". 

Historical commentary: The "options" command was somewhat ported over to GX/cluster-mode from its 7-Mode CLI origin, but there as a strong sentiment to discourage its existence in clustershell. So it was decided to hide it.  Unfortunately, some teams (like NVE/NAE folks) thought it was a good idea to use it for controlling the default. Arrgh.