Update install-mediator-pkg-task.adoc
Field Support had a feedback about the User-Guide:

"Given that the UEFI setting can be a showstopper for installation, I recommend putting the information about UEFI within the steps itself in the docs page.

When the information is in the steps section, it illustrates that the information provided is needed in some fashion for a successful installation. Currently it is in the "About this task" section, which does not properly illustrate the full importance of the UEFI secure boot setting notice."

Hence, making a change to move the "About this task" content as Step 2.