Update index.adoc
Change 'move out of' to 'move over to'.

It currently reads as the user is already in active mode, when the lines preceding this indicate the user is in learning mode. 'move out of' implies the user is already in active mode, but they can't be if they are in learning mode.