Update convert-snapmirror-version-flexible-task.adoc
A customer was badly confused by the cli syntax examples in the published doc:

snapmirror show -destination-path _SVM:volume_|_cluster://SVM/volume_

The pipe separator is used in the CLI reference as well as the command line to denote -or- examples.  In this context, we only use SVM name and volume name.  The example after the pipe is for pre-8.2 relationships like LS Mirror, and has no purpose in the context of DP to XDP conversion.

Additionally, there seems to be a syntax error, prepending and appending underscore to the strings in the syntax example.

The customer encountered an error during an attempted conversion and opened a support case as a result.  The error also led the customer to locate a KB article about errors during DP to XDP conversion, where they also tried to apply the bad syntax.

Once the customer had a remote session with support, and the bad syntax was explained, the steps in the docs as well as steps in the KB article worked.

Please make corrections that align with the simplified SVM:volume examples provided in my edit.